Update README
removing dontrackus because of the EULA 
You are hereby granted a non-exclusive, non-transferable, non-sublicensable limited license to use the Products (in object code form only) for non-commercial uses consistent with the description of the Products we provided with the Products.  This Agreement shall immediately terminate upon any breach of its terms.  Reverse engineering or use of the Products other than as specifically provided herein is strictly prohibited.